Move WeightedEdge out of the MST header into its own
